Rename mongodb.js to database.js
