+++ server/api/tournaments.js
@@ -1,5 +1,5 @@
 const ObjectId = require('mongodb').ObjectId;
-const db = require('../mongodb.js');
+const db = require('../database.js');
 
 module.exports = {
     getAll: (req, res, next) => {
