+++ server/index.js
@@ -4,7 +4,7 @@ const express = require('express');
 const path = require('path');
 const config = require('../tbm.config.js');
 const api = require('./api');
-const mongodb = require('./mongodb.js');
+const mongodb = require('./database.js');
 
 const server = express();
 const publicDir = path.join(__dirname, '..', 'public');
